+++ tests/example/example.spec.js
@@ -1,3 +1,8 @@
+/**
+ * add extra information for file
+ * @owner FO
+ */
+
 const { test, expect } = require('@playwright/test');
 const Util = require('../common/util.js');