+++ tests/home-page/home-page.spec.js
@@ -4,8 +4,6 @@ const EC = require('eight-colors');
 // POM Page Object Model
 const HomePage = require('./home-page.js');
 
-let use;
-
 /**
  * override "beforeAll hook" title to
  * @title do something before all
@@ -13,7 +11,6 @@ let use;
 test.beforeAll(({ browserName }, workerInfo) => {
     console.log(EC.magenta('beforeAll'));
     console.log(`Running ${browserName} in worker #${workerInfo.workerIndex}`);
-    use = test.info().project.use;
 });
 
 /**
@@ -35,7 +32,8 @@ test.afterAll(() => {
 // let pageCurrent;
 
 test('test home page', async ({ page, context }, testInfo) => {
-    const homePage = new HomePage(page, context, use);
+    const metadata = testInfo.config.metadata;
+    const homePage = new HomePage(page, context, metadata);
     await homePage.init();
     await homePage.goto();
     await homePage.checkClientScript();
@@ -61,24 +59,27 @@ test('test home page', async ({ page, context }, testInfo) => {
  * @owner 大强
  * @annotations TODO
  */
-test('test screenshot and video', async ({ page, context }) => {
-    const homePage = new HomePage(page, context, use);
+test('test screenshot and video', async ({ page, context }, testInfo) => {
+    const metadata = testInfo.config.metadata;
+    const homePage = new HomePage(page, context, metadata);
     await homePage.init();
     await homePage.goto();
     await homePage.checkWithError();
 });
 
-test('test without custom steps', async ({ page }) => {
-    await page.goto(use.url);
+test('test without custom steps', async ({ page }, testInfo) => {
+    const metadata = testInfo.config.metadata;
+    await page.goto(metadata.url);
     const locator = page.locator('input[type="search"]');
     await locator.fill('Text content');
     await expect(locator).toHaveValue('Text content');
 });
 
 // https://playwright.dev/docs/api/class-test#test-step
-test('test with custom steps', async ({ page }) => {
+test('test with custom steps', async ({ page }, testInfo) => {
+    const metadata = testInfo.config.metadata;
     await test.step('accessing the login page', async () => {
-        await page.goto(use.url);
+        await page.goto(metadata.url);
     });
     await test.step('completing phone number', async () => {
         const locator = page.locator('input[type="search"]');